+++ natural_selection/__init__.py
@@ -1,5 +1,5 @@
-__version__ = '0.2.27'
-__date__ = "2022-06-29"
+__version__ = '0.2.29'
+__date__ = "2022-11-01"
 
 from time import gmtime
 import multiprocessing as mp
@@ -9,8 +9,13 @@ from datetime import datetime
 import warnings as w
 from typing import Callable, Any, Iterable, List, Union, Type
 import pickle
+import copy
 
 import numpy as np
+
+from natural_selection.utils import get_random_string, clone_classic, default_save_checkpoint_function, \
+    evaluate_individuals_sequentially, evaluate_individual_multiproc_wrapper, IslandError
+from natural_selection.utils.population_growth import population_incremental
 
 from natural_selection.genetic_algorithms import Gene, Chromosome, Individual
 from natural_selection.genetic_programs import Node, GeneticProgram, random_generate
@@ -19,8 +24,8 @@ from natural_selection.genetic_algorithms.operators.initialisation import initia
 from natural_selection.genetic_algorithms.operators.selection import selection_elites_top_n, selection_parents_two, selection_survivors_all
 from natural_selection.genetic_algorithms.operators.crossover import crossover_two_uniform
 from natural_selection.genetic_algorithms.operators.mutation import mutation_randomize
-from natural_selection.genetic_algorithms.utils.probability_functions import crossover_prob_function_classic, mutation_prob_function_classic
-from natural_selection.genetic_algorithms.utils import clone_classic, default_save_checkpoint_function, GeneticAlgorithmError, evaluate_individual
+from natural_selection.utils.probability_functions import crossover_prob_function_classic, mutation_prob_function_classic
+from natural_selection.genetic_algorithms.utils import GeneticAlgorithmError
 
 from natural_selection.genetic_programs.operators.initialisation import initialise_population_full_method
 from natural_selection.genetic_programs.utils import GeneticProgramError
@@ -29,12 +34,13 @@ import natural_selection.genetic_programs.node_operators as gp_func
 
 class Island:
     """
-    A simple Island to perform a Genetic Algorithm. By default the selection, mutation, crossover, and probability functions
-    default to the classic functions.
+    A simple Island to perform a Genetic Algorithm. By default, the selection, mutation, crossover, and probability functions
+    default to the classic functions for Genetic Algorithms.
 
     Args:
         function_params (dict): The parameters for the fitness function (default = None).
         maximise_function (bool): If True, the fitness value is maximised, False will minimise the function fitness (default = True).
+        population_growth_function (Callable): Function that defines what the population growth is, steady, generational etc. (default = None).
         parent_selection (Callable): Function for selecting individuals for crossover (default = None).
         initialisation_function (Callable): A function for randomly creating new individuals from the given adam.
         parent_combination (Callable): Function for combining parents for crossover (default = None).
@@ -53,6 +59,7 @@ class Island:
         filepath (str): If a filepath is specified, the pickled island is loaded from it, skipping the rest of initialisation (default = None).
         save_checkpoint_level (int): Level of checkpoint saving 0 = none, 1 = per generation, 2 = per evaluation (default = 0).
         core_count (int, float): Number of cores to split evaluation on, for all cores, set -1, use float for fractional (default = 1).
+        population_evaluation_function (Callable): Function taking a list of individuals and calling the fitness functions (default = None).
         allow_twins (bool): Only add new offspring to the population if they have a unique chromosome (default = False).
 
     Attributes:
@@ -81,6 +88,7 @@ class Island:
 
     def __init__(self, function_params : dict = None,
                  maximise_function : bool = True,
+                 population_growth_function : Callable = population_incremental,
                  initialisation_function: Callable = initialise_population_random,
                  parent_selection : Callable = selection_elites_top_n,
                  parent_combination : Callable = selection_parents_two,
@@ -99,6 +107,7 @@ class Island:
                  filepath : str = None,
                  save_checkpoint_level : int = 0,
                  core_count : Union[int, float] = 1,
+                 population_evaluation_function : Callable = evaluate_individuals_sequentially,
                  allow_twins : bool = False):
 
         if name is None:
@@ -137,6 +146,8 @@ class Island:
 
         self.maximise_function = maximise_function
         self.verbose_logging(f"island: maximise_function {maximise_function}")
+        self.population_growth_function = population_growth_function
+        self.verbose_logging(f"island: population_growth_function {population_growth_function.__name__}")
         self._initialise = initialisation_function
         self.verbose_logging(f"island: initialisation_function {initialisation_function.__name__}")
         self.parent_selection = parent_selection
@@ -157,7 +168,7 @@ class Island:
         self.verbose_logging(f"island: survivor_selection_function {survivor_selection_function.__name__}")
 
         self.alien_spawn_function = alien_spawn_function
-        self.verbose_logging(f"island: survivor_selection_function {survivor_selection_function.__name__}")
+        self.verbose_logging(f"island: alien_spawn_function {alien_spawn_function.__name__}")
 
         self.save_checkpoint = save_checkpoint_function
         self.verbose_logging(f"island: save_checkpoint_function {save_checkpoint_function.__name__}")
@@ -167,6 +178,9 @@ class Island:
 
         self.random_seed = random_seed
         self.verbose_logging(f"island: random_seed {random_seed}")
+
+        self.population_evaluation_function = population_evaluation_function
+        self.verbose_logging(f"island: population_evaluation_function {population_evaluation_function.__name__}")
 
         if isinstance(core_count, int):
             if core_count < 1:
@@ -184,6 +198,7 @@ class Island:
         self.parents = list()
         self.mutants = list()
         self.children = list()
+        self.deaths = dict()
         self.species_type = "def"
         self.generation_count = 0
         self.save_checkpoint_level = save_checkpoint_level
@@ -232,6 +247,14 @@ class Island:
 
         if save_checkpoint_function and '<lambda>' in repr(save_checkpoint_function):
             w.warn("WARNING: 'save_checkpoint_function' lambda can not be pickled using standard libraries.")
+
+
+        if population_evaluation_function and '<lambda>' in repr(population_evaluation_function):
+            w.warn("WARNING: 'population_evaluation_function' lambda can not be pickled using standard libraries.")
+
+        if population_growth_function and '<lambda>' in repr(population_growth_function):
+            w.warn("WARNING: 'population_growth_function' lambda can not be pickled using standard libraries.")
+
 
         self.__island_properties = dict()
 
@@ -498,6 +521,7 @@ class Island:
                parent_selection_params : dict = None,
                parent_combination_params: dict = None,
                survivor_selection_params: dict = None,
+               population_growth_params: dict = None,
                alien_spawn_params: dict = None,
                criterion_function : Callable = None,
                criterion_params : dict = None,
@@ -515,7 +539,10 @@ class Island:
             mutation_probability (float): Initial mutation probability (default = 0.25).
             crossover_params (dict): Dict of params for custom crossover function (default = None).
             mutation_params (dict): Dict of params for custom mutation function (default = None).
-            selection_params (dict): Dict of params for custom selection function (default = None).
+            parent_selection_params (dict): Dict of params for custom selection function (default = None).
+            parent_combination_params (dict): Dict of params for custom combination function (default = None).
+            survivor_selection_params (dict): Dict of params for custom survivor selection function (default = None).
+            population_growth_params (dict): Dict of params for custom population growth function (defualt = None).
             alien_spawn_params (dict): Dict of params for alien spawn function (default = None).
             criterion_function (Callable): A function to evaluate if the desired criterion has been met (default = None).
             criterion_params (dict): Function parameters for criterion (default = None).
@@ -552,6 +579,11 @@ class Island:
         else:
             _survivor_selection_params = {}
 
+        if population_growth_params:
+            _population_growth_params = population_growth_params
+        else:
+            _population_growth_params = {}
+
         if alien_spawn_params:
             _alien_spawn_params = alien_spawn_params
         else:
@@ -578,6 +610,7 @@ class Island:
                                        parent_selection_params=_parent_selection_params,
                                        parent_combination_params=_parent_combination_params,
                                        survivor_selection_params=_survivor_selection_params,
+                                       population_growth_params=_population_growth_params,
                                        crossover_probability=crossover_probability,
                                        mutation_probability=mutation_probability,
                                        crossover_params=_crossover_params,
@@ -622,13 +655,10 @@ class Island:
 
     def __evaluate_individuals(self, individuals):
         if self.core_count == 1:
-            for individual in individuals:
-                self.verbose_logging(f"eval: {str(individual)}")
-                individual.evaluate(island=self, params=self.function_params)
-            return individuals
+            return self.population_evaluation_function(individuals=individuals, island=self, params=self.function_params)
 
         with mp.Pool(self.core_count) as p:
-            for f, i in zip(p.starmap(evaluate_individual, [(i, self, self.function_params) for i in individuals]), individuals):
+            for f, i in zip(p.starmap(evaluate_individual_multiproc_wrapper, [(i, self, self.function_params) for i in individuals]), individuals):
                 i.fitness = f
 
         return individuals
@@ -638,6 +668,7 @@ class Island:
                               parent_selection_params,
                               parent_combination_params,
                               survivor_selection_params,
+                              population_growth_params,
                               crossover_probability,
                               mutation_probability,
                               crossover_params,
@@ -654,7 +685,7 @@ class Island:
 
         self.parents.append({'generation' : g, 'parents' : selected_parents})
 
-        # Children are strictly copies or new objects seeing as the have a lineage and parents
+        # Children are strictly copies or new objects seeing as they have a lineage and parents
         generation_children = list()
         for parents in self.parent_combination(individuals=selected_parents, island=self, **parent_combination_params):
             self.verbose_logging(f"select: parent_count {len(parents)}")
@@ -678,7 +709,7 @@ class Island:
 
         self.children.append({'generation' : g, 'children' : generation_children})
 
-        # Mutants are not strictly copied but rather only modified seeing as the are part of the children list
+        # Mutants are not strictly copied but rather only modified seeing as they are part of the children list
         generation_mutants = list()
         for mutant in generation_children:
             if self.mutation_prob(mutation_probability=mutation_probability, island=self):
@@ -695,7 +726,9 @@ class Island:
         if self.save_checkpoint_level == 2:
             self.save_checkpoint(event=f'evolve_pre_eval_{g}', island=self)
 
-        for individual in self.__evaluate_individuals(generation_children):
+        generation_children = self.__evaluate_individuals(generation_children)
+
+        for individual in generation_children:
             offspring_fitnesses.append(individual.fitness)
 
         if self.save_checkpoint_level == 2:
@@ -714,13 +747,21 @@ class Island:
                     edge['gene_inheritance'] = gene_inheritance[p.name]
                 self.lineage['edges'].append(edge)
 
+        interim_children = list()
         for individual in self.survivor_selection(individuals=generation_children, island=self, **survivor_selection_params):
             if self.allow_twins or not individual.unique_genetic_code(force_update=True) in self.unique_genome:
                 # If we want a diverse gene pool, this must be true
+                interim_children.append(individual)
                 self.verbose_logging(f"evolve: add {str(individual)}")
-                self.population.append(individual)
                 self.unique_genome.append(individual.unique_genetic_code())
 
+        grown_population, deaths = self.population_growth_function(population=self.population,
+                                                                  offspring=interim_children,
+                                                                  island=self, **population_growth_params)
+        if len(grown_population) == 0:
+            raise IslandError(f"Population length of 0. {len(interim_children)} offspring were created/survived in generation {g}")
+        self.population = grown_population
+        self.deaths[g] = deaths
 
         population_fitnesses = [ind.fitness for ind in self.population]
         selected_parents_fitnesses = [ind.fitness for ind in selected_parents]
@@ -758,6 +799,10 @@ class Island:
 
         self.verbose_logging(f"evolve: stats {self.generation_info[-1]}")
 
+        pop_stats = selection_elites_top_n(island=self,
+                                           individuals=copy.deepcopy(self.population),
+                                           n=len(self.population))
+
         self.generation_info.append(
             {
                 self.__generation_key: g,
@@ -767,14 +812,14 @@ class Island:
                 self.__fitness_std_key: np.std(population_fitnesses),
                 self.__fitness_min_key: min(population_fitnesses),
                 self.__fitness_max_key: max(population_fitnesses),
-                self.__most_fit_key: selection_elites_top_n(island=self, individuals=self.population, n=1)[0].name,
-                self.__least_fit_key: selection_elites_top_n(island=self, individuals=self.population, n=1)[-1].name
+                self.__most_fit_key: pop_stats[0].name,
+                self.__least_fit_key: pop_stats[-1].name
             }
         )
 
         self.verbose_logging(f"evolve: stats {self.generation_info[-1]}")
 
-        elites = selection_elites_top_n(island=self, individuals=self.population, n=4)
+        elites = pop_stats[:4]
         elite_fitnesses = [ind.fitness for ind in elites]
         self.elites.append({'generation': g, 'elites': elites})
 
@@ -797,7 +842,7 @@ class Island:
         for i in self.population:
             i.birthday()
 
-        new_aliens = self.alien_spawn_function(**alien_spawn_params ,island=self)
+        new_aliens = self.alien_spawn_function(**alien_spawn_params, island=self)
 
         if len(new_aliens) > 0:
             alien_fitnesses = list()
@@ -919,21 +964,3 @@ class Island:
         if self.logging_function:
             self.logging_function(event_message=event_message, island=self)
 
-
-def get_random_string(length : int = 8, include_numeric=False) -> str:
-    """
-    Generate a random string with a given length. Used mainly for password generation.
-    Args:
-        length (int): Length to generate.
-        include_numeric (bool): Include numbers?
-
-    Returns:
-        str: Random character string.
-    """
-    import random
-    import string
-
-    letters = string.ascii_letters
-    if include_numeric:
-        letters = f'{letters}{string.digits}'
-    return ''.join(random.choice(letters) for i in range(length))